+++ src/Gravatar.php
@@ -1,41 +1,34 @@
 <?php
-/*
- * This file is part of forxer/Gravatar package.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
 
 namespace Gravatar;
 
+use Gravatar\Concerns\HasEmail;
+
 class Gravatar
 {
-    const URL = '//www.gravatar.com/';
+    use HasEmail;
 
-    /**
-     * @var string The address email to be used.
-     */
-    protected $sEmail;
+    const URL = '//www.gravatar.com/';
 
     /**
      * Return the Gravatar image based on the provided email address.
      *
-     * @param string $sEmail The email to get the gravatar for.
-     * @param string $iSize The avatar size to use, must be less than 2048 and greater than 0.
-     * @param string $sDefaultImage The default image to use. Use a valid image URL, or a recognized gravatar "default".
-     * @param string $sRating The maximum rating to use for avatars
-     * @param string $sExtension The avatar extension to use
-     * @param bool $bForceDefault Force the default image to be always load.
-     * @return \Image
+     * @param string $email The email to get the gravatar for.
+     * @param int|null $size The avatar size to use, must be less than 2048 and greater than 0.
+     * @param string|null $defaultImage The default image to use. Use a valid image URL, or a recognized gravatar "default".
+     * @param string|null $rating The maximum rating to use for avatars
+     * @param string|null $extension The avatar extension to use
+     * @param bool $forceDefault Force the default image to be always load.
+     * @return Image
      */
-    public static function image($sEmail, $iSize = null, $sDefaultImage = null, $sRating = null, $sExtension = null, $bForceDefault = false)
+    public static function image(string $email, ?int $size = null, ?string $defaultImage = null, ?string $rating = null, ?string $extension = null, bool $forceDefault = false): Image
     {
         $gravatarImage = (new Image())
-            ->setEmail($sEmail)
-            ->setSize($iSize)
-            ->setDefaultImage($sDefaultImage, $bForceDefault)
-            ->setMaxRating($sRating)
-            ->setExtension($sExtension);
+            ->setEmail($email)
+            ->setSize($size)
+            ->setDefaultImage($defaultImage, $forceDefault)
+            ->setMaxRating($rating)
+            ->setExtension($extension);
 
         return $gravatarImage;
     }
@@ -43,104 +36,62 @@ class Gravatar
     /**
      * Return multiples Gravatar images based on the provided array of emails addresses.
      *
-     * @param array $aEmail The emails list to get the Gravatar images for.
-     * @param string $iSize The avatar size to use, must be less than 2048 and greater than 0.
-     * @param string $sDefaultImage The default image to use. Use a valid image URL, or a recognized gravatar "default".
-     * @param string $sRating The maximum rating to use for avatars.
-     * @param string $sExtension The avatar extension to use.
-     * @param bool $bForceDefault Force the default image to be always load.
+     * @param array $emails The emails list to get the Gravatar images for.
+     * @param int|null $size The avatar size to use, must be less than 2048 and greater than 0.
+     * @param string|null $defaultImage The default image to use. Use a valid image URL, or a recognized gravatar "default".
+     * @param string|null $rating The maximum rating to use for avatars
+     * @param string|null $extension The avatar extension to use
+     * @param bool $forceDefault Force the default image to be always load.
      * @return array
      */
-    public static function images(array $aEmail, $iSize = null, $sDefaultImage = null, $sRating = null, $sExtension = null, $bForceDefault = false)
+    public static function images(array $emails, ?int $size = null, ?string $defaultImage = null, ?string $rating = null, ?string $extension = null, bool $forceDefault = false): array
     {
-        $aImages = [];
+        $images = [];
 
-        foreach ($aEmail as $sEmail) {
-            $gravatarImage = (new Image())
-                ->setEmail($sEmail)
-                ->setSize($iSize)
-                ->setDefaultImage($sDefaultImage, $bForceDefault)
-                ->setMaxRating($sRating)
-                ->setExtension($sExtension);
-
-            $aImages[$sEmail] = $gravatarImage;
+        foreach ($emails as $email) {
+            $images[$email] = (new Image())
+                ->setEmail($email)
+                ->setSize($size)
+                ->setDefaultImage($defaultImage, $forceDefault)
+                ->setMaxRating($rating)
+                ->setExtension($extension);
         }
 
-        return $aImages;
+        return $images;
     }
 
     /**
      * Return the Gravatar profile URL based on the provided email address.
      *
-     * @param string $sEmail The email to get the Gravatar profile for.
-     * @param string $sFormat The profile format to use.
-     * @return \Profile
+     * @param string $email The email to get the Gravatar profile for.
+     * @param string|null $sFormat The profile format to use.
+     * @return Profile
      */
-    public static function profile($sEmail, $sFormat = null)
+    public static function profile(string $email, ?string $format = null): Profile
     {
         return (new Profile())
-            ->setEmail($sEmail)
-            ->setFormat($sFormat);
+            ->setEmail($email)
+            ->setFormat($format);
     }
 
     /**
      * Return multiples Gravatar profiles based on the provided array of emails addresses.
      *
      * @param array $aEmail The emails list to get the Gravatar profiles for.
-     * @param string $sFormat The profile format to use.
+     * @param string|null $sFormat The profile format to use.
      * @return array
      */
-    public static function profiles(array $aEmail, $sFormat = null)
+    public static function profiles(array $emails, ?string $format = null): array
     {
-        $aProfils = [];
+        $profils = [];
 
-        foreach ($aEmail as $sEmail) {
-            $gravatarProfil = (new Profile())
-                ->setEmail($sEmail)
-                ->setFormat($sFormat);
-
-            $aProfils[$sEmail] = $gravatarProfil;
+        foreach ($emails as $email) {
+            $profils[$email] = (new Profile())
+                ->setEmail($email)
+                ->setFormat($format);
         }
 
-        return $aProfils;
-    }
-
-    /**
-     * Get or set the address email to be used.
-     *
-     * @param string $sEmail
-     * @return number|\Gravatar
-     */
-    public function email($sEmail = null)
-    {
-        if (null === $sEmail) {
-            return $this->getEmail();
-        }
-
-        return $this->setEmail($sEmail);
-    }
-
-    /**
-     * Get the address email to be used.
-     *
-     * @return int The current avatar size in use.
-     */
-    public function getEmail()
-    {
-        return $this->sEmail;
-    }
-
-    /**
-     * Set the address email to be used.
-     *
-     * @param string $sEmail
-     * @return \Gravatar
-     */
-    public function setEmail($sEmail)
-    {
-        $this->sEmail = $sEmail;
-
-        return $this;
+        return $profils;
     }
 
     /**
@@ -149,7 +100,7 @@ class Gravatar
      * @param string $sEmail The email to get the hash for.
      * @return string The hashed form of the email.
      */
-    protected static function getHash($sEmail)
+    protected static function hash($sEmail)
     {
         return md5(strtolower(trim($sEmail)));
     }